Refactor finance strategy classes and optimize price fetching
Stream operations in CsvPriceRepositoryImpl have been optimized and refactored for better performance. Also, renamed the BuyAndHoldVariablePositionImplFullDataStdDev strategy class to AbstractBuyAndHoldVariablePositionStrategy and extracted the calculation logic for contracts to hold in a separate ContractsToHoldCalculator class. Unnecessary loggings in various strategies were removed and restoration of a commented strategy run in BackTest was done.